Special :starred command to bring up starred logs
This may not be the best idea, but its a start. The reason there is
a setTimeout there is so that first the LogViewer gets cleared out
before new logs get put in it. When I didn't have the setTimeout
there, it would render the logs that were part of the last search
result.