Removed Time Cursor and stopped creating bins
The time cursor was being updated often (on scroll) causing the
whole chart to redraw all the time. I'll move it into it's own
svg component so that other things don't need to redraw.

As for the bins, I was creating bins that were different from the
server bins. Now I calculate the number of bins only and use whatever
ts keys are given to me by the server.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch blank-histogram
# Changes to be committed:
#	modified:   src/js/components/CountByTime.js
#	modified:   src/js/reducers/countByTime.js
#	modified:   src/js/reducers/countByTime.test.js
#
# Changes not staged for commit:
#	modified:   src/js/receivers/countByTimeReceiver.js
#	modified:   src/js/receivers/eventsReceiver.js
#