Include count() in log subset queries
When brushing the histogram, the search stats at the bottom were
only showing that 1000 events matched. This was accurate, but
confusing to the user since they do not see that we tack on "head 1000"
to the end of their query. Therefore, add count() so that all the
data is in fact touched and the stats are accurate.