Add reducer for log viewer component
It will keep track of things like when its fetching more logs,
when it has reached the end of the logs, the startts and stopts
of the visible logs and maybe more.

It will not contain the array of logs. I want to save this log viewer
information but not save the actual logs.