Flow LogViewer and combine connector with component
Now that we add types to the props, it makes sense to me that the
prop types and the stateToProps should be right next to each other
in the same file. The connectors folder should be deprecated. It
adds too much indirection and duplication. Keep the components small
and hook them up to the state they need right there in the file.
Then export default the normal non-connected class, and export the
XName of the class by name.