Use selected log instead of index
I wanted to use the index for performance, but when I open another
log in the detail view without clicking on the row, I don't know
the index. Now we just compare the objects and see if they are the same.
Computers are fast.

Although this might now work if the log is the same but we performed
another search and now the reference is different.