refactor: Better mock client
Have the MockApi extend from the real api. Then only override the
send method, since all the api calls use the send message. It's
an excellent place to override so that we don't make the http
request.

In the future I could see myself giving the mock api the ability
to return specified responses in a clear way.