chore: use separate config for itest
A separate configuration for itest, and removing itest from the main
jest "roots", allows for "npm test" to work how it previously did.
"npm run itest" also still works, by finding the configuration in
itest/config.json. Jest documentation says the root is implied as the
directory of the config.

https://jestjs.io/docs/en/cli#config-path