fix: No more races
I seperated out the concept of aborting a request into two ideas.

The first is manually killing a request. This should run the
"aborted" callback.

The second is issuing a search before the previous search was
finished. This should not run the aborted callback.

This solves alot of the headaches.