test: histogram deep inspection
This test inspects the rendered histogram to a reasonable but not
precise degree. It depends on corelight data being loaded. It also
depends on current product behavior. It verifies that a number of rect
elements exist, and that their attributes look reasonable. It also
verifies that the rect elements vary by _path, which is handled by g
elements of varying classes.