test: wait for all elements PROD-437
When I was writing PoC tests, I was being lazy and not waiting for all
elements. But PROD-437 shows even login page elements may not be ready.
This patch uses waitForExist as part of the builtin webdriver API to
wait for elements to exist. I had missed this before when I wrote
retry(), so new element waiting can be implemented with this. I ensured
the failpath works with clear error messages by writing typo'd
selectors.