feat: Serialization framework
Take care of saving things to local state in one place.
Now we can save date objects in the store, persist them
to local storage, and then recreate the date objects when
we pull them out of storage. There is also a framework for
adding more custom types that we can serialize/deserialize.