test: ensure that 'reset state' in-app works (PROD-397)
This test attempts to catch bugs like PROD-515 and PROD-539. We run a
search and then reset the state of the app. At that point, we look for
an empty login page, and if we receive one, we log in again. We finally
ensure that the search input is empty again.