refactor: "libify" login form elements
When the earliest tests were written they used hardcoded elements, which
at the time I knew was wrong, but I wanted to test the end-to-end
Spectron more than I cared about good test design. Now that I'm back in
the area of the login form, refactor those selectors to use the new
library-based way, in which we put a marker in the product.