fix: Bug fixes with pinned programs
Serveral bugs were found and this commit addresses them. One thing
to note is that we must always treat the "previous" search bar
value as the one to stick on the end of the main program, not the
current. The current value may be something else because we use
it to edit pins. We must rename "current" to "draft" and "previous"
to "head" or something.