refactor: Chart rendering is simple now
Each chart object has all the data it needs to mount, draw, and
update itself. Each chart object has an array of "pens" which
contain a mount() fn and a draw() fn. I'm not sure if they should
be housed on the chart object or seperate, but on the chart object
seems nice for now. This is another preparation for a more chart
intesive UI coming down the road.