test fix: make search results deterministic
Data in different logs will be ingested in different order. This means a
connection could appear in the conn log "first" and the path-specific
log "second" in some cases, or vice versa. Fix a bug in which we were
searching on a non-conn path (dns) and sometimes getting the wrong cell.

Fix another bug in which a search could produce non-determinstic
results. The fix is to sort on both ts and uid.