add gsutil capabilities to integration tests image (PROD-802)
In order to load meaningfully larger data sets for the integration
tests, change the integration test CI's image. Use the image that the
correctness tests have been using for a couple months. That image is the
existing integration test image already in use with gsutil installed on
top. The effective diff is here:

https://github.com/looky-cloud/rug/commit/e53cb3f30a31272b6e4866a2a98234ece724183f#diff-bebe822d715123ec31b03d977212fe2a