integration tests: do not clobber chromedriver logs
To get more insight into PROD-853 we must stop clobbering chromedriver
logs. (My local testing with tail -f assumed we wouldn't clobber, but
it's clear we do.) The fastest way not to clobber is to tell
newAppInstance() a way to set a unique path for each test's log. Do that
using the __filename under test and a test ID number.

This is a hack that takes on some tech debt but the idea is to get the
logs asap. I can refactor test sections and other things later.