test feat: add a test search with a somewhat complicated AST
The search is:

_path=x509 | every 1d count() by certificate.version | sort ts, certification.version

This uses a search term, every, a group by, and sort on two columns.
It's a search used to test viewing the AST, but I'd be more confident
knowing that this search is also producing correct results. That way
when the underlying AST changes, we can be confident in changing its
snapshot as long as the results don't change.

I verified the results match the equivalent Postgres query results.