feat: A very hand useSearch hook
I'd like to move all our searches in the app to use this hook.
Rather than having all the searches centralized and stored in
the redux store, keep the local to the component that needs it.

I even think the main search could use this.

I'd like to then keep a small list of the past 50 searches and
some stats about them to use in the search inspector.