feat: Add a flows file for tabs state
The flows functions are for conditional or complex state changes,
where you need to look at the state to determine what to do.

I think it makes sense to put then in their state objects if they
are only concerned with that slice of state. More complex state
maniuplaton goes in the flows top level directory.