handle null durations for pcap downloads
If a conn log is missing the duration field, brim throws an error
and hangs with a notice of "Downloading" at the bottom of the screen.

This commit hacks in a workaround to return 0 for the second
an nanosecond accessors on a log field.  This will be fixed more
holisticly in a later PR.

There is lots of ugliness to get tests to pass.  Something is
definitely wrong here and I invite help to get this right.