feat: Show loading message after selecting pcaps
I removed the modal version of the progress indicator for a simple
spinner on the page. It will be displayed until the pcap has been
processed and is entirely ready to search.

Soon zqd will provide streaming updates, in which case we will add
an indicator somewhere else on the page with the progress.