download zeek and make available for zqd (#417)
Download the zeek artifact created by the brimsec/zeek release process, and launch zqd with a path that includes the zeek executable. This will let zqd spawn zeek as needed for the pcap conversion process.

The actual zeek that zqd launches will be a script that prepares the needed Zeek environment variables, then execs the necessary binary.