feat: Convert Zng types back to Zeek Types
The app has knowledge of zeek types and formats types of data
appropriately. Since zng has adopted its own type system, we need
to now convert those types back to zeek. This is not an ideal
solution but one that we can do quickly. The way we alias types
is in flux and may change soon in a way that permits us to remove
this code.