PROD-1505 Confine pcap ingest to its original tab (#454)
* fix: Allow an optional tabId for actions related to search

The problem addressed was: Ingest a pcap on one tab. Before it
finishes, switch to another tab. When the ingest finishes, it set
the space name and time frame on whichever tab was "active" at the
time it finished, not at the time it started.

To address this, we add the optional tabId parameter to the actions
that change the search parameters (space and span). There is already
logic in Tabs.reducer that checks for the presence of an arg
called tabId in the passed in action. If it is present, it uses it.
If not, it uses whatever the active tab is at that moment.

We cache the tab that was open when ingest began, then set the
search parameters on that tab when ingest finishes.

* fix: Order of actions once ingest has finished

After the ingest is finisehd:
  * Fetch the updated space data from zqd
  * Set space in the search params for that tab
  * Set the progress to 100%
  * Wait 1.5 sec for the transition
  * Set the progress to null (meaning no longer ingesting)