PROD-1589 Handle TaskEnd errors from zqd (#473)
* feat: Handle the optional TaskEnd errors from zqd

* refactor: Brim Error Objects

Instead of using app error classes, use simple js objects with the
keys "type", "message", and "details". This captures all the
functionality of the AppError class and is serializable. It also
allows for creating a specific error object, which was impossible
to do with the AppErrors.

To keep this diff as small as possible, keep the AppError classes
but just treat them like functions to create BrimError objects.

* feat: Add a pcapIngest error object

* fix: lint error

* fix: tests

* refactor: use lodash upperFirst