fix: Spaces.getSpaces(clusterId) bug fix
Because of the way the spaces are stored with the name as the key
and the details as the value, when we get the spaces, we were
getting undefined because spaces had not had their details loaded
yet.

This commit changes the selector to always merge in the name (key)
of the space into the details object so that we don't get
undefined values in the spaces array.