feat: Extend Handlers reducer for Ingest Requests
This splits out the handlers into two types: SEARCH and INGEST

The SEARCH types have an abort function to abort the search,
but the INGEST types do not at the moment. They will be used to
determine if the confirm dialog should appear when a user closes
the window.

The INGEST handlers also contain the spaceName so that if the user
does decide to close the window and cancel the ingest, we can
delete the space.