feat: Move onBeforeUnload to its own flow
In this function, we get the spaces that are still ingesting in
this window. If there are more than zero, ask the user if they
really want to close the window and explain what will happen if
they do.

If they confirm, then run the shutdown flow. It deletes the
partial spaces, persists the state, and destroys the window.

Destroy does not call the onbeforeunload callback like "close".

Using destroy here prevents infinite confirmation dialog boxes.