feat: Use the new onBeforeUnload function
This is in a setTimeout because the async nature of the dialog
box messed up the e.returnValue call. The window would close
every time no matter what the user selected.

So what we do is we schedule the onBeforeUnload flow for the next
tick around the event loop, then prevent the window fom closing
right away. The logic in the other file will handle the destuction
of the window if the user confirms.