PROD-1635 Warn on close if ingesting (#499)
* fix: Spaces.getSpaces(clusterId) bug fix

Because of the way the spaces are stored with the name as the key
and the details as the value, when we get the spaces, we were
getting undefined because spaces had not had their details loaded
yet.

This commit changes the selector to always merge in the name (key)
of the space into the details object so that we don't get
undefined values in the spaces array.

* feat: Add a function to destroy the window

* feat: Extend Handlers reducer for Ingest Requests

This splits out the handlers into two types: SEARCH and INGEST

The SEARCH types have an abort function to abort the search,
but the INGEST types do not at the moment. They will be used to
determine if the confirm dialog should appear when a user closes
the window.

The INGEST handlers also contain the spaceName so that if the user
does decide to close the window and cancel the ingest, we can
delete the space.

* fix: Update existing use of Handler.register

To account for the changes in the commit above

* feat: Register an ingest handler when we open a pcap

It would be great to make a hook in the zealot client that
emits all requests that it makes. That way we could register and
unregister all the requests that zealot ever makes from a single
place. For now, this works.

* feat: Move onBeforeUnload to its own flow

In this function, we get the spaces that are still ingesting in
this window. If there are more than zero, ask the user if they
really want to close the window and explain what will happen if
they do.

If they confirm, then run the shutdown flow. It deletes the
partial spaces, persists the state, and destroys the window.

Destroy does not call the onbeforeunload callback like "close".

Using destroy here prevents infinite confirmation dialog boxes.

* feat: Use the new onBeforeUnload function

This is in a setTimeout because the async nature of the dialog
box messed up the e.returnValue call. The window would close
every time no matter what the user selected.

So what we do is we schedule the onBeforeUnload flow for the next
tick around the event loop, then prevent the window fom closing
right away. The logic in the other file will handle the destuction
of the window if the user confirms.

* fix: Always refresh the spaces when the app starts up

* fix: Check if window is destroyed before sending

* fix: Allow app to quit even if warnings popup

* fix: Split close from reload

Delegate all the closing logic to the renderer process which will
call destroy() if necessary.

Split up the concept of reloading and closing into two different
flows.