fix: Use the search.ts to know when things changed
Things meaning the search params (span, program). The Span controls
show a little blue dot when the state is "dirty". Meaning the value
has changed, but the use has not submitted the search.

Same for the "RUN" button. It needs to know when the search bar is
in a dirty state to appear or not.