PROD-1637 Search bar dirty form state (#502)
* fix: Split history push into local and global

Investigation.push is global
History.push is local to the tab

* feat: Add ts to the Search reducer

This ts represents the timestamp of the last time the search bar
was submitted, whether it was saved to history or not.

* fix: Use the search.ts to know when things changed

Things meaning the search params (span, program). The Span controls
show a little blue dot when the state is "dirty". Meaning the value
has changed, but the use has not submitted the search.

Same for the "RUN" button. It needs to know when the search bar is
in a dirty state to appear or not.

* fix: Add regenerator-runtime

And sorry my editor auto-arranges the imports