test improvements (#561)
* test fix: deflake external API tests

- Ensure jest won't time out a test.
- Use retry() for the zeek.org tests.

* ci: control xvfb where needed

The screen size provided by xvfb-action isn't adequate for our tests.
This sets the screen size for when we need XVFB. The tradeoff is a YAML
"fork" depending on OS. It's likely that down the road OS-wise
integration tests will vary anyway, so I think some sort of fork is
going to happen sooner or later. It is a reasonable tradeoff.