test fix: test download pcaps (fixes #543) (#637)
Add integration tests for downloading packet captures. The tests are
accomplished by uploading data, running zql to find a deterministic
connection (can't use UID), clicking a cell, and downloading the packet
capture. In CI the files are put into a directory, where we can examine
the file's MD5.

In older times, there was a bug where a null-duration connection's pcap
could not be downloaded. That has since been fixed, so that test is
enabled.