ci fix: properly handle windows paths in bash
Adjust how the npm cache directory is expanded. We need to quote it in
bash so that "\" is not quashed. But it's not possible to directly quote
the Github action expression, because the running shell will attempt to
expand it using normal shell expansion rules. That means this doesn't
work:

  - run: rm -rf "${{ foo.bar }}"

Also,

  - run: rm -rf ${{ foo.bar }}

will not work in some cases, like if foo.bar is a Windows path. If
foo.bar "c:\foo", the shell will turn that into "c:foo".

The fix is to assign an environment variable from the expression
directly, then quote the environment variable in the shell.

To force using a new, smaller cache, change the cache name.