feat: auto-update zq after successful testing
When a zq PR is merged, update zq locally, run unit and integration
tests, and commit the change if the tests pass. If the tests fail, save
artifacts and open a PR. The PR text will contain links to the failing
auto-update attempt. Opening a PR lets us piggy-back on alert
infrastructure already in place, and provides a branch on which further
Brim adjustments could be made.