feat: tron.session
The session class now runs any pending migrations on data that
it loads in from a file.

Session also is now only concerned with saving and loading data
that its given. Formating the state for our Brim App is now
done in formatSessionState.js

Another change is that there will only be one instance of this
class in the whole app. It will be like a singleton so that we
can keep the current version in the closure, and update it
after we migrate.