zq update through "archive metadata with ordered import info" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/795, authored by @alfred-landrum,
has been merged.

archive metadata with ordered import info

(on top of https://github.com/brimsec/zq/pull/787)

These changes cause import to enforce an ordering on incoming data, and have it remember the mapping of spans to zng files in the zar.json metadata file.  The 'mkdirs' command is removed, since the zar directories will be created as needed during archive walks.