zq update through "zapi: Signal based context" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/809, authored by @mattnibs,
has been merged.

zapi: Signal based context

For zapi add a context to the root command that cancels when
SIGTERM or SIGINT is recieved. This context should be passed into
any long running tasks launched by the command.

This functionality will come more into play when the longer running
commands such search or post data are shortly introduced.