Migration session data (#787)
* feat: tron.migrations

A class to load all the files in the migration folder, determine
which migrations are pending, run them, and update the version.

* feat: tron.session

The session class now runs any pending migrations on data that
it loads in from a file.

Session also is now only concerned with saving and loading data
that its given. Formating the state for our Brim App is now
done in formatSessionState.js

Another change is that there will only be one instance of this
class in the whole app. It will be like a singleton so that we
can keep the current version in the closure, and update it
after we migrate.

* feat: Use new classes in main.js

* feat: windowManager.fetchWindowStates

From the main process, this function asks all open search windows
to send back their state. It them waits 5 seconds for them to
reply on a one time use channel. Once all windows have sent the
state over, we resolve the promise.

I moved the area where we decided what is persistable into
the getPersistable() function.

I also exteded the test/setup.js file to mock out more methods
on BrowserWindow.

* feat: Use fetchWindowStates before saving the session

The state that "Save Session for Testing Migrations" would save
was previously state. The windowManager used to only get updates
about the window state right before they closed.

Adding this function here ensures that the windowManager has the
most recent state from all open windows.

* fix: Correctly type file.contents

* fix: Popuplate migrations folder

* cleanup: Remove old ways of versioning state

* feat: Add migration for new zeek runner pref

* fix: Use true current version not latest migration version

* fix: Handle case where state is undefined

* fix: lint