zq update through "Page iterator" by aswan
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/813, authored by @aswan,
has been merged.

Page iterator

This is definitely a first cut but the desire on Slack was to create a PR for it now.  I have not had a chance to clean this up, some specific things that need cleaning are:
- error handling is sloppy and inconsistent, it many cases we just panic when encountering some parquet feature or variant that is not yet handled
- various bits of duplicated code
- little attention has been given to comments, names of functions/variables, ordering of functions in the source filed, etc.
- the wide reader interfaces that have methods for every parquet primitive type make for very verbose code.  @mccanne 's suggestion for this is to push the zng encoding all the way into the iterators which would also improve performance.

I think the last item would be best done in a separate follow-up.  As for the rest, I'm aware of all the cleanup that needs to happen, a review with detailed comments about those issues is probably not a good use of anybody's time at this stage...