zq update through "Stop using ParquetReader completely" by aswan
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/825, authored by @aswan,
has been merged.

Stop using ParquetReader completely

The main thing we were still using the ParquetReader class for was its
SchemaHandler.  But that was clumsy as it still had vestiges of the
renaming-columns-to-valid-go-public-field-names gunk that was needed for
its reflect-based api.

In this update, things that care about the parquet schema read the parquet
file footer directly, bypassing parqet-go's SchemaHandler class.  We're
still using the ParquetFile interface and low-level handlers for thrift,
compression, etc. from parquet-go, but we're now totally free from its
reader package.