zq update through "Let expr tests be written with nil records" by aswan
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/835, authored by @aswan,
has been merged.

Let expr tests be written with nil records

Every expression is evaluated in the context of a zng record, but a bunch
of our tests just evaluate literal expressions so requiring that they
create a record makes the tests more verbose and harder to follow.  Fix
the test helpers to allow for a nil record and clean up affected tests.