zq update through "zar index: add -i flag to select input file" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/829, authored by @henridf,
has been merged.

zar index: add -i flag to select input file

This PR adds the `-i` option to `zar index`, allowing it to operate over other files than the usual archive log file.

With this change, the one remaining `zar zdx` example in the demo can be moved to `zar index` (which landed while brimsec/zq#816 was ongoing).