zq update through "initial archive index search api" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/833, authored by @alfred-landrum,
has been merged.

initial archive index search api

This adds an api endpoint for index queries of the form:
```
POST /space/{space_id}/indexsearch
{
  "index_name": "<name>"
  "patterns": ["<pattern1>",...]
}
```

Valid requests are similar to the options for `zar find`: For "standard" indices, requests have a blank index_name, and a single pattern of forms like `:ip=10.0.0.1` or `uri=/hello`.  Custom indices require specifying an index name with one or more patterns; there's not yet an api to list the available indices for an archive space.

Note: As a side effect of making this work on Windows, the file names for zar indices have changed, replacing colons with dashes, eg: `zdx:type:ip.zng` becomes `zdx-type-ip.zng`. Colons are not valid filename characters for Windows.