add a test that creates a zar space under brim
Brim isn't fully ready to handle zar spaces yet, but this test
establishes a temporary workflow for getting zar data into a space and
making sure Brim starts. As Brim's functionality to read a Zar space
improves, this test can be modified. The purpose of this is to put
something in place that changes alongside the functionality as opposed
to having nothing until all the functionality is ready. As such, there's
not as much abstraction as there otherwise would be: much of this should
get thrown away once Brim supports Zar.

The test uses Brim as the driver for starting zqd so that the test don't
have to do zqd orchestration itself. It ingests sample.tsv as usual.
Then it uses zealot to create a second, empty space called sample.zar
and also get space metadata from space sample.tsv.brim. The test then
imports samle.tsv.brim's all.zng into sample.zar, with a small enough
chop size to ensure multiple files. The test creates a number of
indexes. Finally, it restarts Brim and ensures Brim restarts.