zq update through "detector.MultiFileReader" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/863, authored by @mattnibs,
has been merged.

detector.MultiFileReader

MultiReader returns a zbuf.ReadCloser that's the logical concatenation of the
provided input paths. They're read sequentially. Once a stream has reached EOF,
Close is called on the file. Once all inputs have reached end of stream,
Read will return end of stream. If any of the readers return a non-nil error,
Read will return that error.

The primary use case is for reading large amounts of files but
limiting the amount of open file descriptors.

Use MultiFileReader when doing zar searches in zqd as well as
performing searches via the zar command.

closes brimsec/zq#832