zq update through "support archive based subspaces" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/854, authored by @alfred-landrum,
has been merged.

support archive based subspaces

Add apis to support archive backed spaces and subspaces.

The space creation api is extended to explicitly support archive backed spaces. A new space can point to an existing archive root, or can specify the creation of a new archive. However, there's not yet support for an archive space to reflect changes made via a separate `zar import` to the same root, but that's easy and will come shortly.

The api to create a subspace of an archive store backed space is:
```
POST /space/{parent_space_id}/subspace
{
"name": "<name>",
"open_options": {
  "log_filter": ["logid1", ...]
  },
}
```
The response is the same as the existing space post request. The subspace supports index and zql searches, and changing its name, via the existing apis.

Closes brimsec/zq#590