zq update through "Link to Zeek From Home video" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/883, authored by @philrz,
has been merged.

Link to Zeek From Home video

Since we covered some `zar` basics in the recent Zeek From Home session, here I'm adding a link directly to that portion of the video from the top of the README. This may be helpful for readers who are willing to "lean back" and learn but may lack the motivation to "lean forward" and execute the commands themselves.