zq update through "Flip words to match modified example" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/894, authored by @philrz,
has been merged.

Flip words to match modified example

After talking to @henridf about the modified example in brimsec/zq#893, I realized at the last minute that the descriptive wording with the example should change to match.