zq update through "Decomposable reducers" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/892, authored by @henridf,
has been merged.

Decomposable reducers

Add a `ConsumePart()` and `ResultPart()` methods and implement them in reducers that perform "decomposable aggregations". And a couple of supporting commits before that.

Decomposition isn't used yet (other than in the test), but putting it up as a self-contained standalone change on the journey to spillable groupby.