zq update through "zq: return error of writer.Close" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/902, authored by @mattnibs,
has been merged.

zq: return error of writer.Close

Because of the way s3manager is written, we may only find out about
an s3 error until after Writer.Close is called. The zq command was
ignoring writer.Close errors. Do not do this.

Closes brimsec/zq#900