zq update through "Fix "every" entry in v0.15.0 changelog" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/909, authored by @philrz,
has been merged.

Fix "every" entry in v0.15.0 changelog

I made a mistake when compiling the `v0.15.0` changelog, copying `-every` as shown in the opening text of https://github.com/brimsec/zq/issues/828 rather than plain `every` which is the actual syntax. Thanks to @nwt for spotting this. I've already made the correction in the "live" notes on the Releases tab and on my announcement on public Slack, so, doubling back to fix it here.

While I'm at it, another mistake I noticed: I'd accidentally linked to the Issue rather than the PR, which goes against convention, so I've corrected that here.