fix: Show generic message
Since we no longer have the space name readily available, change
all the tooltips to simply say "This space no longer exists".

Also remove the notice that appears on a tab when the space
no longer exists. It was too noisy and caused confusion when you
haven't opened the app in a while.