Remove space name (#875)
* refactor: Sync zqd spaces state in the initializer

* refactor: Remove the spaces from the global store

* refactor: Do not use spaceName in the tab bar

* refactor: Remove spaceName from Search.setSpace

* fix: Remove references to spaceName

* fix rebase

* fix: Show generic message

Since we no longer have the space name readily available, change
all the tooltips to simply say "This space no longer exists".

Also remove the notice that appears on a tab when the space
no longer exists. It was too noisy and caused confusion when you
haven't opened the app in a while.