zq update through "Zeek compatibility overhaul" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/919, authored by @philrz,
has been merged.

Zeek compatibility overhaul

As of the opening of this PR, the version of this doc in `master` is mostly remnants from the ZSON days when we were attempting to maintain true backward compatibility with the Zeek TSV format. Therefore it was due for an overhaul to explain current behaviors.

This represents my attempt to rationalize what I observe as the black box treatment of Zeek data as ZNG and how it's written back out again as Zeek by `zq`. Reviewers who know the code are encouraged to contribute better rationalizations, as well as pointing out cases I might have forgotten to include.

Also, if you feel any of the behaviors documented here are undesirable to the point where we wish they did not occur, I'd be open to pausing to revise the ZNG spec and `zq` implementation such that the Zeek compatibility story reads better. The special treatment of `_path` as `string` might be one such area, but personally, I'm also fine leaving the behavior as-is and letting the doc cover the topic.

Finally, https://github.com/brimsec/zq/pull/917 should ideally be merged before this one, since we refer to `enum` here, but the spec is currently slightly inaccurate in terms of how it describes the existing (non-) support for a true ZNG `enum`.