adjust advance zq ci timeouts (#888)
* adjust advance zq ci timeouts

Webdriver-style tests wait for elements to appear or to be set to a
specific state. This means a lot of test failures are timeout-driven.
An overall job timeout is already set here in case turnstyle hangs.
However, the current value is too short in cases where Zq and Brim are
so broken as to cause every integration test to fail. Locally under
these conditions, the integration tests took 63 minutes to fail. In CI,
assume that could take up to 90 minutes. Set that as the integration
test step timeout. Increase the job timeout to 120.

Other PRs will make other adjustments and hopefully allow for these
numbers to be reduced or removed.

* use --bail 10