zq update through "Spillable groupby" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/932, authored by @henridf,
has been merged.

Spillable groupby

When the groupby table grows to a certain size (defined by the existing `-limit` parameter), if reducers are decomposable, spill to disk rather than giving up. Spills are sorted and then merged when reading.

closes brimsec/zq#800