zq update through "More ZNG spec fixes" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/935, authored by @philrz,
has been merged.

More ZNG spec fixes

After polishing the [Zeek compatibility](https://github.com/brimsec/zq/blob/master/zng/docs/zeek-compat.md) doc, I came back to the ZNG spec yet again with intent to see if more de-Zeek-ification was now possible, and I ended up finding more fundamental things to scrub, including removal of a couple remaining relics from ZSON days that don't seem easy to resuscitate. Other than fixes for correctness, the most significant changes are in the spirit of trying to have the spec be in sync with the current `zq` implementation... _or_ calling out explicitly where the two are known to not be in sync.