zq update through "Publish ES Module ZQL" by jameskerr
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/978, authored by @jameskerr,
has been merged.

Publish ES Module ZQL

In addition to publishing a commonjs version of zql.js for use with Node runtimes, also publish an ES module version. This will make it easy to import into the Deno zealot client, or directly into a modern browser.

This conversion is done using the rollup node module.