zq update through "Fix alias bug with context resets" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/964, authored by @henridf,
has been merged.

Fix alias bug with context resets

The fix here is to use the "value" of the alias type (consisting of the alias name and aliased type ID) rather than the alias ID when building a record key. This avoids the issue where alias IDs for the same alias could be different in different batches. (The underlying problem seems to be that the Cocke and Schwartz algorithm isn't being applied to alias IDs... but I digress).

closes brimsec/zq#866