zq update through "Check key type consistency in custom indexes" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/991, authored by @henridf,
has been merged.

Check key type consistency in custom indexes

closes brimsec/zq#815

While in there I noticed we didn't have a test for multi-key custom indexes and so added one in the first commit. The second commit has the actual fix and a test verifying the fix.