zq update through "ztest: return compilation error as string" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1001, authored by @henridf,
has been merged.

ztest: return compilation error as string

Previously, compilation errors weren't returned as the "warnOrError"
return argument of ztest.runzq(), and so a zql-style test couldn't
check for those. This change allows zql-style ztests to check for
compilation errors.