zq update through "Index metadata" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1003, authored by @henridf,
has been merged.

Index metadata

This fulfills the requirements of brimsec/zq#989 (I think) in a very simple way, but putting this up as a draft to see if this is what @alfred-landrum had in mind.

In particular, the the spec says "use the entries in the metadata file to choose what index to use for lookup." which could probably be interpreted in other ways. Very open to revising this if needed!