zq update through "address more cases where typecontext should be provided" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1009, authored by @alfred-landrum,
has been merged.

address more cases where typecontext should be provided

The case in search.go is the only one that looks like it could lead to immediate trouble, but I went ahead and changed the rest of the Storage methods as well.