zq update through "Make s3io.Reader an io.Seeker" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1011, authored by @mattnibs,
has been merged.

Make s3io.Reader an io.Seeker

This functionality is needed for the zar find command to work.

Add system test for zar index and zar find commands.

Part of brimsec/zq#972 & brimsec/zq#973